requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.87.0
h11==0.14.0
httptools==0.5.0
idna==3.4
ImageHash==4.3.1
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.5
Pillow==9.3.0
pydantic==1.10.2
python-dotenv==0.21.0
python-multipart==0.0.5
PyWavelets==1.4.1
PyYAML==6.0
scipy==1.9.3
six==1.16.0
sniffio==1.3.0
starlette==0.21.0
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
